Address comments
- Pull CsvExportable into separate concern
- Give background job default priority
