+++ app/jobs/generate_output_csv_job.rb
@@ -3,6 +3,10 @@
 class GenerateOutputCsvJob < ApplicationJob
   queue_as :default
 
+  def default_priority
+    -100
+  end
+
   def perform(batch_process)
     batch_process.output_csv
   rescue => e
