+++ app/models/batch_process.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 class BatchProcess < ApplicationRecord # rubocop:disable Metrics/ClassLength
+  include CsvExportable
   attr_reader :file
   after_create :determine_background_jobs
   before_create :mets_oid
@@ -49,27 +50,6 @@ class BatchProcess < ApplicationRecord # rubocop:disable Metrics/ClassLength
   def oids
     return @oids ||= parsed_csv.entries.map { |r| r['oid'] } unless csv.nil?
     @oids ||= [oid]
-  end
-
-  def output_csv
-    return nil unless batch_action == "export child oids"
-    headers = ["child_oid", "parent_oid", "order", "parent_title", "label", "caption", "viewing_hint"]
-    csv_string = CSV.generate do |csv|
-      csv << headers
-      oids.each do |oid|
-        begin
-          po = ParentObject.find(oid.to_i)
-          po.child_objects.each do |co|
-            row = [co.oid, po.oid, co.order, po.authoritative_json["title"]&.first, co.label, co.caption, co.viewing_hint]
-            csv << row
-          end
-        rescue ActiveRecord::RecordNotFound
-          row = ["----", oid, "", "Parent Not Found in database", "", ""]
-          csv << row
-        end
-      end
-      csv_string
-    end
   end
 
   def created_file_name
